Revert "Create Hourly Forecast View"
