+++ WeatherApp/Info.plist
@@ -21,5 +21,9 @@
 			</array>
 		</dict>
 	</dict>
+	<key>NSLocationAlwaysUsageDescription</key>
+	<string>This app uses your location in order to provide you accurate weather information.</string>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>This app uses your location in order to provide you accurate weather information.</string>
 </dict>
 </plist>
