+++ WeatherApp/SceneDelegate.swift
@@ -11,8 +11,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
     let viewModel = ViewModel()
-    let viewController = ViewController()
-    let heroView = HeroView()
     
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
@@ -29,6 +27,13 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
 
     func sceneDidBecomeActive(_ scene: UIScene) {
+        Task {
+            do {
+                try await viewModel.getWeatherForUserLocation()
+            } catch {
+                print(error)
+            }
+        }
     }
 
     func sceneWillResignActive(_ scene: UIScene) {
@@ -37,25 +42,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
 
     func sceneWillEnterForeground(_ scene: UIScene) {
-        Task {
-            do {
-                try await viewModel.getWeatherForUserLocation()
-            } catch {
-                print(error)
-            }
-        }
-        
-        // Does it really updates?
-        
-        viewModel.currentWeatherModel.bind(fire: true) { currentWeatherModel in
-            if let currentWeatherModel = currentWeatherModel {
-                DispatchQueue.main.async {
-                    self.heroView.update(with: currentWeatherModel)
-                }
-                //it is visible on the screen?
-            }
-        }
-        
         // Called as the scene transitions from the background to the foreground.
         // Use this method to undo the changes made on entering the background.
     }
