+++ WeatherApp/Services/WeatherForecastModel.swift
@@ -11,6 +11,7 @@ struct WeatherForecastModel: Codable {
     var cnt: Int
     var list: [WeatherForHour]
     var city: CityDetails
+   
 }
 
 struct WeatherForHour: Codable {
